# backend-rust/spec/golden-openapi.json
{
...
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/internal/jobs/{jobId}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "internal-job-controller"
        ]
      }
    },
...
    "/api/jobs/clear": {
      "delete": {
        "operationId": "clearQueue",
        "parameters": [
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "job-controller"
        ]
      }
    },
...
    "/api/jobs/{id}": {
      "delete": {
        "operationId": "deleteJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "job-controller"
        ]
      }
    },
    "/api/jobs/{id}/pause": {
      "post": {
        "operationId": "pauseJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "job-controller"
        ]
      }
    },
    "/api/jobs/{id}/resume": {
      "post": {
        "operationId": "resumeJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "job-controller"
        ]
      }
    },
    "/api/jobs/{id}/retry": {
      "post": {
        "operationId": "retryJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "job-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080/tlhub"
    }
  ],
...
}